+++ packages/react/src/ActionList/Group.tsx
@@ -4,12 +4,29 @@ import Box from '../Box'
 import type {SxProp} from '../sx'
 import {ListContext, type ActionListProps} from './shared'
 import type {AriaRole} from '../utils/types'
-import {default as Heading} from '../Heading'
 import type {ActionListHeadingProps} from './Heading'
 import {useSlots} from '../hooks/useSlots'
 import {defaultSxProp} from '../utils/defaultSxProp'
 import {invariant} from '../utils/invariant'
 import {clsx} from 'clsx'
+import {useFeatureFlag} from '../FeatureFlags'
+import classes from './ActionList.module.css'
+import groupClasses from './Group.module.css'
+
+type HeadingProps = {
+  as?: 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6'
+  className?: string
+  children: React.ReactNode
+} & SxProp
+
+const Heading: React.FC<HeadingProps> = ({as: Component = 'h3', className, children, sx = defaultSxProp}) => {
+  return (
+    // Box is temporary to support lingering sx usage
+    <Box as={Component} className={className} sx={sx}>
+      {children}
+    </Box>
+  )
+}
 
 export type ActionListGroupProps = {
   /**
@@ -50,9 +67,10 @@ export const Group: React.FC<React.PropsWithChildren<ActionListGroupProps>> = ({
   auxiliaryText,
   selectionVariant,
   role,
-  sx = {},
+  sx = defaultSxProp,
   ...props
 }) => {
+  const enabled = useFeatureFlag('primer_react_css_modules_team')
   const id = useId()
   const {role: listRole} = React.useContext(ListContext)
 
@@ -72,6 +90,55 @@ export const Group: React.FC<React.PropsWithChildren<ActionListGroupProps>> = ({
     groupHeadingId = id
   }
 
+  if (enabled) {
+    if (sx !== defaultSxProp) {
+      return (
+        <Box as="li" className={groupClasses.Group} role={listRole ? 'none' : undefined} sx={sx} {...props}>
+          <GroupContext.Provider value={{selectionVariant, groupHeadingId}}>
+            {title && !slots.groupHeading ? (
+              // Escape hatch: supports old API <ActionList.Group title="group title"> in a non breaking way
+              <GroupHeading variant={variant} auxiliaryText={auxiliaryText} _internalBackwardCompatibleTitle={title} />
+            ) : null}
+            {/* Supports new API ActionList.GroupHeading */}
+            {!title && slots.groupHeading ? React.cloneElement(slots.groupHeading) : null}
+            <ul
+              // if listRole is set (listbox or menu), we don't label the list with the groupHeadingId
+              // because the heading is hidden from the accessibility tree and only used for presentation role.
+              // We will instead use aria-label to label the list. See a line below.
+              aria-labelledby={listRole ? undefined : groupHeadingId}
+              aria-label={listRole ? title ?? (slots.groupHeading?.props.children as string) : undefined}
+              role={role || (listRole && 'group')}
+            >
+              {slots.groupHeading ? childrenWithoutSlots : props.children}
+            </ul>
+          </GroupContext.Provider>
+        </Box>
+      )
+    }
+    return (
+      <li className={groupClasses.Group} role={listRole ? 'none' : undefined} {...props}>
+        <GroupContext.Provider value={{selectionVariant, groupHeadingId}}>
+          {title && !slots.groupHeading ? (
+            // Escape hatch: supports old API <ActionList.Group title="group title"> in a non breaking way
+            <GroupHeading variant={variant} auxiliaryText={auxiliaryText} _internalBackwardCompatibleTitle={title} />
+          ) : null}
+          {/* Supports new API ActionList.GroupHeading */}
+          {!title && slots.groupHeading ? React.cloneElement(slots.groupHeading) : null}
+          <ul
+            // if listRole is set (listbox or menu), we don't label the list with the groupHeadingId
+            // because the heading is hidden from the accessibility tree and only used for presentation role.
+            // We will instead use aria-label to label the list. See a line below.
+            aria-labelledby={listRole ? undefined : groupHeadingId}
+            aria-label={listRole ? title ?? (slots.groupHeading?.props.children as string) : undefined}
+            role={role || (listRole && 'group')}
+            className={groupClasses.GroupListTemp}
+          >
+            {slots.groupHeading ? childrenWithoutSlots : props.children}
+          </ul>
+        </GroupContext.Provider>
+      </li>
+    )
+  }
   return (
     <Box
       as="li"
@@ -97,7 +164,7 @@ export const Group: React.FC<React.PropsWithChildren<ActionListGroupProps>> = ({
           // because the heading is hidden from the accessibility tree and only used for presentation role.
           // We will instead use aria-label to label the list. See a line below.
           aria-labelledby={listRole ? undefined : groupHeadingId}
-          aria-label={listRole ? (title ?? (slots.groupHeading?.props.children as string)) : undefined}
+          aria-label={listRole ? title ?? (slots.groupHeading?.props.children as string) : undefined}
           role={role || (listRole && 'group')}
         >
           {slots.groupHeading ? childrenWithoutSlots : props.children}
@@ -113,6 +180,7 @@ export type ActionListGroupHeadingProps = Pick<ActionListGroupProps, 'variant' |
   React.HTMLAttributes<HTMLElement> & {
     as?: 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6'
     _internalBackwardCompatibleTitle?: string
+    variant?: 'filled' | 'subtle'
   }
 
 /**
@@ -125,7 +193,7 @@ export type ActionListGroupHeadingProps = Pick<ActionListGroupProps, 'variant' |
  */
 export const GroupHeading: React.FC<React.PropsWithChildren<ActionListGroupHeadingProps>> = ({
   as,
-  variant,
+  variant = 'subtle',
   // We are not recommending this prop to be used, it should only be used internally for incremental rollout.
   _internalBackwardCompatibleTitle,
   auxiliaryText,
@@ -134,7 +202,7 @@ export const GroupHeading: React.FC<React.PropsWithChildren<ActionListGroupHeadi
   sx = defaultSxProp,
   ...props
 }) => {
-  const {variant: listVariant, role: listRole} = React.useContext(ListContext)
+  const {role: listRole} = React.useContext(ListContext)
   const {groupHeadingId} = React.useContext(GroupContext)
   // for list role, the headings are proper heading tags, for menu and listbox, they are just representational and divs
   const missingAsForList = (listRole === undefined || listRole === 'list') && children !== undefined && as === undefined
@@ -152,58 +220,47 @@ export const GroupHeading: React.FC<React.PropsWithChildren<ActionListGroupHeadi
     `Looks like you are trying to set a heading level to a ${listRole} role. Group headings for ${listRole} type action lists are for representational purposes, and rendered as divs. Therefore they don't need a heading level.`,
   )
 
-  const styles = {
-    paddingY: '6px',
-    paddingX: listVariant === 'full' ? 2 : 3,
-    fontSize: 0,
-    fontWeight: 'bold',
-    color: 'fg.muted',
-    ...(variant === 'filled' && {
-      backgroundColor: 'canvas.subtle',
-      marginX: 0,
-      marginBottom: 2,
-      borderTop: '1px solid',
-      borderBottom: '1px solid',
-      borderColor: 'neutral.muted',
-    }),
-
-    [`.ActionListGroupHeading`]: {
-      fontSize: 'var(--text-body-size-small), 12px',
-      fontWeight: 'var(--base-text-weight-semibold, 600)',
-      lineHeight: 'var(--text-body-lineHeight-small, 1.6666)',
-      color: 'var(--fgColor-muted)',
-    },
-
-    [`.ActionListGroupHeadingDescription`]: {
-      fontSize: 'var(--text-body-size-small, 12px)',
-      fontWeight: 'var(--base-text-weight-normal, 400)',
-      lineHeight: 'var(--text-body-lineHeight-small, 1.6666)',
-      color: 'var(--fgColor-muted)',
-    },
-  }
-
   return (
     <>
       {/* for listbox (SelectPanel) and menu (ActionMenu) roles, group titles are presentational. */}
       {listRole && listRole !== 'list' ? (
-        <Box sx={styles} role="presentation" aria-hidden="true" {...props}>
-          <span id={groupHeadingId}>{_internalBackwardCompatibleTitle ?? children}</span>
-          {auxiliaryText && <div className="ActionListGroupHeadingDescription">{auxiliaryText}</div>}
-        </Box>
+        <div
+          role="presentation"
+          className={groupClasses.GroupHeadingWrap}
+          aria-hidden="true"
+          {...props}
+          data-variant={variant}
+        >
+          <span className={clsx(className, groupClasses.GroupHeading)} id={groupHeadingId}>
+            {_internalBackwardCompatibleTitle ?? children}
+          </span>
+          {auxiliaryText && <div className={classes.Description}>{auxiliaryText}</div>}
+        </div>
       ) : (
         // for explicit (role="list" is passed as prop) and implicit list roles (ActionList ins rendered as list by default), group titles are proper heading tags.
-        <Box sx={styles}>
-          <Heading
-            className={clsx(className, 'ActionListGroupHeading')}
-            as={as || 'h3'}
-            id={groupHeadingId}
-            sx={sx}
-            {...props}
-          >
-            {_internalBackwardCompatibleTitle ?? children}
-          </Heading>
-          {auxiliaryText && <div className="ActionListGroupHeadingDescription">{auxiliaryText}</div>}
-        </Box>
+        <div className={groupClasses.GroupHeadingWrap} data-variant={variant}>
+          {sx !== defaultSxProp ? (
+            <Heading
+              className={clsx(className, groupClasses.GroupHeading)}
+              as={as || 'h3'}
+              id={groupHeadingId}
+              sx={sx}
+              {...props}
+            >
+              {_internalBackwardCompatibleTitle ?? children}
+            </Heading>
+          ) : (
+            <Heading
+              className={clsx(className, groupClasses.GroupHeading)}
+              as={as || 'h3'}
+              id={groupHeadingId}
+              {...props}
+            >
+              {_internalBackwardCompatibleTitle ?? children}
+            </Heading>
+          )}
+          {auxiliaryText && <div className={classes.Description}>{auxiliaryText}</div>}
+        </div>
       )}
     </>
   )
